Accept @Header annotation values that are objects
toString is called on the supplied values, then used.

I included this in one of my projects that was failing, and it worked as intended.

Felt like there should've been a more end-to-end test that I could've included here, but I didn't easily see it, and I'm a bit short on time, so point it out if you feel like that needs to be added as well.