Allow passing in an executor on which callback methods are invoked.
This only has an effect when Call is the return type of a service method. This does not affect the callback methods of a Call passed to a CallAdapter.

We also change the non-Android call adapter factory to be a simple implementation that returns the original instance rather than needlessly wrapping it in the executor factory with a synchronous executor.