Add a Kotlin coroutine adapter.
This supports returning a `Deferred` from your service methods allowing an easy call to `await()` inside a suspending function.

Kotlin coroutines are experimental. This does not make them API unstable, however, which is essential for including them inside Retrofit. When non-experimental coroutines are released we will add a second adapter for them and mark this one as deprecated for removal in the next major version. For more information see https://stackoverflow.com/a/46240340/132047.