Break request creation parsing out of service method
The parsing of annotations which define the shape of a request and the mechnism of request creation is independent of the response converter and call adapter parsing in the return type. Moreover, the request creation part is common to things like web sockets and server-sent events which might be supported in the future where the return type handling is completely different.