Implement correct isDefault check for Android
There is no ability to invoke a default method yet, but at least now you'll get a UOE thrown instead of it being treated as a normal service method.

Unfortunately we can't test this for the same reason we don't test normal Java 8 functionality, it requires the language level be set to 8 and we can't vary the value between the main sources and test sources.