Use per-method locking when parsing annotations
Historically we used a single lock for the whole map. This caused high contention in startup scenarios where multiple threads invoke different methods for the first time and require parsing.

Prior to this we changed to a service-level lock, but this only mitigates the problem for those with partitioned service interfaces. It is overwhelmingly common to only have a single service interface.

We cannot use Map.computeIfAbsent yet because it requires minSdk 24 (and we are 21).

Until then, map values now go through a two-phase lifecycle to provide granular, per-method locking. First, we atomically insert a lock object into the map where successful insertion grants you the right to do parsing. Other threads which discover the lock object will wait to gain the lock. The locking thread will update the map with the parsed model and release the lock. Other threads may now take the lock and discover the parsed model. Callers coming in later never see a lock object, only the parsed model.